Extract upsell to sidebar-layout.js
