+++ packages/js/src/general/routes/alert-center.js
@@ -1,20 +1,11 @@
-import { useSelect } from "@wordpress/data";
 import { __ } from "@wordpress/i18n";
 import { Paper, Title } from "@yoast/ui-library";
-import { PremiumUpsellList } from "../../shared-admin/components";
 import { Notifications, Problems } from "../components";
-import { STORE_NAME } from "../constants";
-import { useSelectGeneralPage } from "../hooks";
 
 /**
  * @returns {JSX.Element} The general page content placeholder.
  */
 export const AlertCenter = () => {
-	const isPremium = useSelectGeneralPage( "selectPreference", [], "isPremium" );
-	const premiumLinkList = useSelectGeneralPage( "selectLink", [], "https://yoa.st/17h" );
-	const premiumUpsellConfig = useSelectGeneralPage( "selectUpsellSettingsAsProps" );
-	const { isPromotionActive } = useSelect( STORE_NAME );
-
 	return (
 		<>
 			<Paper className="yst-p-8 yst-grow">
@@ -25,17 +16,10 @@ export const AlertCenter = () => {
 					</p>
 				</header>
 			</Paper>
-			<div className="yst-grid yst-grid-cols-1 @3xl:yst-grid-cols-2 yst-gap-6 yst-my-6 yst-grow yst-items-start">
+			<div className="yst-grid yst-grid-cols-1 @3xl:yst-grid-cols-2 yst-gap-6 yst-mt-6 yst-grow yst-items-start">
 				<Problems />
 				<Notifications />
 			</div>
-			{ isPremium ? null : (
-				<PremiumUpsellList
-					premiumLink={ premiumLinkList }
-					premiumUpsellConfig={ premiumUpsellConfig }
-					isPromotionActive={ isPromotionActive }
-				/>
-			) }
 		</>
 	);
 };
